+++ app/src/main/java/com/innovatrics/dot/samples/faceautocapture/FaceAutoCaptureResult.kt
@@ -2,7 +2,6 @@ package com.innovatrics.dot.samples.faceautocapture
 
 import android.graphics.Bitmap
 import com.innovatrics.dot.face.quality.FaceAspects
-import com.innovatrics.dot.face.quality.FaceAttribute
 import com.innovatrics.dot.face.quality.FaceQuality
 
 data class FaceAutoCaptureResult(
@@ -10,5 +9,4 @@ data class FaceAutoCaptureResult(
     val confidence: Double?,
     val faceAspects: FaceAspects?,
     val faceQuality: FaceQuality?,
-    val passiveLivenessFaceAttribute: FaceAttribute?,
 )